refactor queue

+++ Contents/Code/bookmark_storage.py
@@ -6,22 +6,13 @@ class BookmarkStorage(FileStorage):
     def __init__(self, file_name):
         FileStorage.__init__(self, file_name)
 
-    def sanitize(self, item):
-        new_item = MediaInfo(item.type)
-
-        for key, value in item.iteritems():
-            if item[key]:
-                new_item[key] = value
-
-        return new_item
-
     def find(self, search_item):
-        search_item = self.sanitize(search_item)
+        BookmarkStorage.sanitize(search_item)
 
         found = None
 
         for item in self.data:
-            mode = search_item.type()
+            mode = search_item.type
 
             if item['path'] == search_item['path']:
                 if mode == MediaInfo.SIMPLE:
@@ -55,6 +46,9 @@ class BookmarkStorage(FileStorage):
             self.save()
 
     def remove(self, item):
-        Storage.remove(self, item)
+        bookmark = self.find(item)
 
-        self.save()
+        if bookmark:
+            Storage.remove(self, item)
+
+            self.save()
